% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ovml_cache_dir}
\alias{ovml_cache_dir}
\title{Path to the cache directory used for model weight files and other data}
\usage{
ovml_cache_dir()
}
\value{
The path as a string
}
\description{
Path to the cache directory used for model weight files and other data
}
