% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo.R
\name{ovml_class_labels}
\alias{ovml_class_labels}
\title{Class labels}
\usage{
ovml_class_labels(dataset = "coco")
}
\arguments{
\item{dataset}{string: which dataset?}
}
\value{
A character vector of class labels
}
\description{
Class labels
}
