% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{ovml_example_image}
\alias{ovml_example_image}
\title{Example images provided as part of the ovideo package}
\usage{
ovml_example_image(choices = 1)
}
\arguments{
\item{choices}{integer: which image files to return?
\itemize{
\item 1 - an image from a match between GKS Katowice and MKS Bedzin during the 2018/19 Polish Plus Liga
\item 2 - the standard YOLO dog image
}}
}
\value{
Path to the image files
}
\description{
Example images provided as part of the ovideo package
}
