% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo_onnx.R
\name{ovml_install_onnxruntime}
\alias{ovml_install_onnxruntime}
\title{Install Python dependencies for ONNX inference}
\usage{
ovml_install_onnxruntime(gpu = FALSE)
}
\description{
Install Python dependencies for ONNX inference
}
