% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo_onnx.R
\name{ovml_onnx_inspect}
\alias{ovml_onnx_inspect}
\title{Inspect ONNX model input/output names and shapes}
\usage{
ovml_onnx_inspect(onnx_file)
}
\description{
Inspect ONNX model input/output names and shapes
}
