% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ovml_plot}
\alias{ovml_plot}
\alias{ovml_ggplot}
\title{Preview plot of detections over image using base graphics}
\usage{
ovml_plot(
  img,
  detections,
  line_args = list(col = "blue", lwd = 1),
  label_args = list(col = "white", cex = 0.75)
)

ovml_ggplot(
  img,
  detections,
  line_args = list(col = "blue", size = 0.75, fill = NA),
  label_args = list(col = "white", size = 2.5, fill = "blue"),
  label_geom = "label"
)
}
\arguments{
\item{img}{string or image: filename of jpg image, or image as read by [\code{\link[jpeg:readJPEG]{jpeg::readJPEG()}}]}

\item{detections}{data.frame: as returned by \code{\link[=ovml_yolo_detect]{ovml_yolo_detect()}}}

\item{line_args}{list: parameters passed to \link{lines} (for \code{\link[=ovml_plot]{ovml_plot()}}) or [\code{\link[ggplot2:geom_tile]{ggplot2::geom_rect()}}] (for \code{\link[=ovml_ggplot]{ovml_ggplot()}})}

\item{label_args}{list: parameters passed to \link{text}}

\item{label_geom}{string: for \code{\link[=ovml_ggplot]{ovml_ggplot()}}, the geom function to use for labels. Either "text" (use [\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}]) or "label" ([\code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}])}
}
\description{
Preview plot of detections over image using base graphics
}
\examples{

## define some demo data
dets <- data.frame(class = rep("person", 3),
                   score = rep(0.99, 3),
                   xmin = c(829, 611, 736),
                   xmax = c(960, 733, 836),
                   ymin = c(88, 258, 213),
                   ymax = c(278, 444, 385),
                   stringsAsFactors = FALSE)
img <- system.file("extdata/images/2019_03_01-KATS-BEDS-frame.jpg", package = "ovml")
ovml_plot(img, dets, line_args = list(col = "red", lwd = 2))
ovml_ggplot(img, dets) + ggplot2::theme_void()

}
