% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo.R
\name{ovml_yolo}
\alias{ovml_yolo}
\title{Construct YOLO network}
\usage{
ovml_yolo(version = 4, device = "cpu", weights_file = "auto")
}
\arguments{
\item{version}{integer or string: currently 3 (YOLO v3), 4 (YOLO v4), or "4-tiny" (YOLO v4-tiny)}

\item{device}{string: "cpu" or "cuda"}

\item{weights_file}{string: either the path to the weights file that already exists on your system or "auto". If "auto", the weights file will be downloaded if necessary and stored in the directory given by \link{ovml_cache_dir}}
}
\value{
A YOLO network object

\dontrun{
  dn <- ovml_yolo()
  img <- ovml_example_image()
  res <- ovml_yolo_detect(dn, img)
  ovml_plot(img, res)
}
}
\description{
Construct YOLO network
}
\references{
https://github.com/pjreddie/darknet
}
