% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo.R
\name{ovml_yolo_detect.default}
\alias{ovml_yolo_detect.default}
\title{Detect objects in image using a YOLO network}
\usage{
\method{ovml_yolo_detect}{default}(net, ...)
}
\description{
Detect objects in image using a YOLO network
}
