% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{ovml_example_image}
\alias{ovml_plot}
\alias{ovml_ggplot}
\alias{ovml_cache_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{ovml.common}{\code{\link[ovml.common]{ovml_cache_dir}}, \code{\link[ovml.common]{ovml_example_image}}, \code{\link[ovml.common:ovml_plot]{ovml_ggplot}}, \code{\link[ovml.common]{ovml_plot}}}
}}

