% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolov3.R
\name{ovml_yolo}
\alias{ovml_yolo}
\title{Construct YOLO network}
\usage{
ovml_yolo(version = 3, device = "cpu", weights_file = "auto")
}
\arguments{
\item{version}{integer: currently only 3, for YOLO v3}

\item{device}{string: "cpu" or "cuda"}

\item{weights_file}{string: either the path to the weights file that already exists on your system or "auto". If "auto", the weights file will be downloaded if necessary and stored in the directory given by \link{ovml_cache_dir}}
}
\value{
A YOLO network object
}
\description{
Construct YOLO network
}
\references{
https://github.com/pjreddie/darknet
}
